Add listener.

+++ base/sys/inc/eventTrigger.h
@@ -5,7 +5,7 @@
 
 namespace parrot
 {
-enum class eIoAction : uint16_t;
+enum class eIoAction : uint8_t;
 
 class EventTrigger : public IoEvent
 {
